requirements.txt
# ===== CORE DEPENDENCIES =====
# Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Data Processing
numpy==1.24.3
pandas==2.2.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0

# ===== OPTIONAL (for development) =====
# Testing
pytest==8.0.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0

# Logging
loguru==0.7.2
